Rename to filename to plural and update references
