+++ src/helpers/logBreadcrumbs.test.js
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from 'vitest';
 import { addBreadcrumb } from '@sentry/vue';
-import { logBreadcrumb, createAuthBreadcrumb, logNavBreadcrumb } from './logBreadcrumb';
+import { logBreadcrumb, createAuthBreadcrumb, logNavBreadcrumb } from './logBreadcrumbs';
 
 vi.mock('@sentry/vue', () => ({
   addBreadcrumb: vi.fn(),
