+++ src/pages/CleverLanding.vue
@@ -12,7 +12,7 @@ import _get from 'lodash/get';
 import _union from 'lodash/union';
 import AppSpinner from '@/components/AppSpinner.vue';
 import { fetchDocById } from '@/helpers/query/utils';
-import { createAuthBreadcrumb, logNavBreadcrumb } from '@/helpers/logBreadcrumb';
+import { createAuthBreadcrumb, logNavBreadcrumb } from '@/helpers/logBreadcrumbs';
 
 const router = useRouter();
 const authStore = useAuthStore();
