+++ src/pages/SSOAuthPage.vue
@@ -14,7 +14,7 @@ import useSSOAccountReadinessVerification from '@/composables/useSSOAccountReadi
 import AppSpinner from '@/components/AppSpinner.vue';
 import { APP_ROUTES } from '@/constants/routes';
 import { AUTH_SSO_PROVIDERS } from '../constants/auth';
-import { logNavBreadcrumb } from '@/helpers/logBreadcrumb';
+import { logNavBreadcrumb } from '@/helpers/logBreadcrumbs';
 
 const router = useRouter();
 const authStore = useAuthStore();
